MakerKeRO/requirements.txt
wget
tgcrypto
ffmpeg-python
py-tgcalls==0.9.7
telethon==1.35.0
pyrogram==2.0.106
youtube-search
youtube-search-python
gitpython
git+https://github.com/yt-dlp/yt-dlp.git
git+https://github.com/alexmercerind/youtube-search-python
git+https://github.com/erdewit/nest_asyncio.git
git+https://github.com/AsmSafone/SafoneAPI
python-dotenv
dnspython
asgiref
asSQL
lyricsgenius
numpy
unidecode
getids
aiofiles
dragonxxdlib
SpeechRecognition
aiohttp
requests
pillow==9.5.0
heroku3
psutil
future
speedtest-cli
motor
pyromod~=3.1.6
googletrans
pytz
tornado==6.4.1
